// app/src/components/Dashboard/Types.ts
export type BorrowItem = {
  quarter: any;
  year: any;
  id: number;
  name: string;
  position: string;
  department: string;
  subDepartment: string;
  phone: string;
  equipment: string;
  lateDays: number;

  requestDate: string;
  borrower: string;
  deviceCategory: string;
  deviceCode: string;
  quantity: number;
  reason: string;
  location: string;
  operator: string;
  borrowDate: string;
};